api_request_app/app/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.6
requests==2.24.0
schedule==0.6.0
urllib3==1.25.10
